fix event/category postman tests

+++ ewm-service/src/main/java/ru/practicum/ewm/event/controller/EventPublicController.java
@@ -16,7 +16,7 @@ public class EventPublicController {
 
     @GetMapping
     public List<EventFullDto> getRequiredPublicEvents(@RequestParam(required = false) String text,
-                                                      @RequestParam(required = false) List<Integer> categories,
+                                                      @RequestParam(required = false) List<Long> categories,
                                                       @RequestParam(required = false) Boolean paid,
                                                       @RequestParam(required = false) String rangeStart,
                                                       @RequestParam(required = false) String rangeEnd,
