+++ ewm-service/src/main/java/ru/practicum/ewm/event/service/EventAdminService.java
@@ -2,16 +2,14 @@ package ru.practicum.ewm.event.service;
 
 import ru.practicum.ewm.event.dto.EventFullDto;
 import ru.practicum.ewm.event.dto.EventUpdateByAdminDto;
-import ru.practicum.ewm.event.dto.EventUpdateByUserDto;
 
 import java.util.List;
 
 public interface EventAdminService {
 
-    // админ может отправлять запрос на изменение события, с одним из двух статусов: PUBLISH_EVENT, REJECT_EVENT
     EventFullDto updateEventByAdmin(Long eventId, EventUpdateByAdminDto eventUpdateByAdminDto);
 
-    List<EventFullDto> getRequiredAdminEvents(List<Integer> users, List<String> states, List<Integer> categories,
+    List<EventFullDto> getRequiredAdminEvents(List<Long> users, List<String> states, List<Long> categories,
                                               String rangeStart, String rangeEnd, Integer from, Integer size);
 
 }
